refactor: semerrgroup wraps errgroup with a semaphore
