+++ pipeline/checksums/checksums.go
@@ -8,12 +8,11 @@ import (
 	"path/filepath"
 
 	"github.com/apex/log"
-	"golang.org/x/sync/errgroup"
 
 	"github.com/goreleaser/goreleaser/checksum"
 	"github.com/goreleaser/goreleaser/context"
 	"github.com/goreleaser/goreleaser/internal/artifact"
-	"github.com/goreleaser/goreleaser/internal/semaphore"
+	"github.com/goreleaser/goreleaser/internal/semerrgroup"
 	"github.com/goreleaser/goreleaser/internal/tmpl"
 )
 
@@ -48,8 +47,7 @@ func (Pipe) Run(ctx *context.Context) (err error) {
 	}
 	defer file.Close() // nolint: errcheck
 
-	var g errgroup.Group
-	var sem = semaphore.New(ctx.Parallelism)
+	var g = semerrgroup.New(ctx.Parallelism)
 	for _, artifact := range ctx.Artifacts.Filter(
 		artifact.Or(
 			artifact.ByType(artifact.UploadableArchive),
@@ -57,10 +55,8 @@ func (Pipe) Run(ctx *context.Context) (err error) {
 			artifact.ByType(artifact.LinuxPackage),
 		),
 	).List() {
-		sem.Acquire()
 		artifact := artifact
 		g.Go(func() error {
-			defer sem.Release()
 			return checksums(file, artifact)
 		})
 	}
